--- a/CLAUDE.md
+++ b/CLAUDE.md
@@ -99,3 +99,6 @@ Pipeline      System      Factory
 - Establish real-time communication channels for sprint coordination
 - Set up monitoring and alerting systems for immediate issue detection
 - Create sprint retrospective and post-mortem documentation template
+
+## Recent Memories
+- 21done