--- a/src/torematrix/core/operations/type_management/__init__.py
+++ b/src/torematrix/core/operations/type_management/__init__.py
@@ -1,56 +1,71 @@
 """Type Management Operations
 
-This package provides comprehensive type management operations including:
-- Bulk type changes and modifications
-- Type conversions with data preservation  
+Agent 3 implementation for bulk operations and management including:
+- Bulk type changes with progress tracking
+- Safe type conversions with data preservation 
 - Type migrations and versioning
-- Batch processing for large-scale operations
-- Progress tracking and rollback capabilities
-- Performance optimization for enterprise workloads
+- Rollback and undo capabilities
+- Performance optimization for large-scale operations
 """
 
-from .bulk_operations import BulkTypeOperationEngine, BulkOperationOptions, BulkOperationResult
-from .conversions import TypeConversionEngine, ConversionResult, ConversionAnalysis
-from .migrations import TypeMigrationManager, MigrationResult, MigrationPlan
-from .warnings import ConversionWarningSystem, WarningLevel, ConversionWarning
-from .batch_processor import BatchProcessor, BatchResult, BatchOptions
-from .progress import ProgressTracker, OperationProgress, ProgressCallback
-from .rollback import RollbackManager, RollbackOperation, RollbackState
-from .optimization import PerformanceOptimizer, OptimizationStrategy, OptimizationResult
+from .bulk_operations import (
+    BulkTypeOperationEngine, BulkOperationOptions, BulkOperationResult,
+    BulkChangePreview, ElementChange, OperationStatus, ConflictResolution
+)
+from .conversions import (
+    TypeConversionEngine, ConversionRule, ConversionAnalysis, ConversionResult,
+    ConversionStrategy, ConversionRisk, DataMappingRule
+)
+from .progress import (
+    ProgressTracker, OperationProgress, ProgressPhase, ProgressCallback,
+    start_operation_progress, update_operation_progress, complete_operation_progress,
+    subscribe_to_operation_progress, get_progress_tracker
+)
+from .rollback import (
+    RollbackManager, RollbackOperation, RollbackAction, RollbackResult,
+    RollbackState, OperationType
+)
 
 __all__ = [
-    # Core engines
+    # Bulk Operations
     'BulkTypeOperationEngine',
-    'TypeConversionEngine', 
-    'TypeMigrationManager',
-    'ConversionWarningSystem',
-    'BatchProcessor',
-    'ProgressTracker',
-    'RollbackManager',
-    'PerformanceOptimizer',
-    
-    # Data classes
-    'BulkOperationOptions',
+    'BulkOperationOptions', 
     'BulkOperationResult',
-    'ConversionResult',
-    'ConversionAnalysis',
-    'MigrationResult',
-    'MigrationPlan',
-    'ConversionWarning',
-    'BatchResult',
-    'BatchOptions',
-    'OperationProgress',
-    'RollbackOperation',
-    'RollbackState',
-    'OptimizationResult',
+    'BulkChangePreview',
+    'ElementChange',
+    'OperationStatus',
+    'ConflictResolution',
     
-    # Enums
-    'WarningLevel',
+    # Type Conversions
+    'TypeConversionEngine',
+    'ConversionRule',
+    'ConversionAnalysis', 
+    'ConversionResult',
+    'ConversionStrategy',
+    'ConversionRisk',
+    'DataMappingRule',
+    
+    # Progress Tracking
+    'ProgressTracker',
+    'OperationProgress',
+    'ProgressPhase',
     'ProgressCallback',
-    'OptimizationStrategy',
+    'start_operation_progress',
+    'update_operation_progress',
+    'complete_operation_progress',
+    'subscribe_to_operation_progress',
+    'get_progress_tracker',
+    
+    # Rollback System
+    'RollbackManager',
+    'RollbackOperation',
+    'RollbackAction',
+    'RollbackResult',
+    'RollbackState',
+    'OperationType',
 ]
 
 # Version info
 __version__ = "1.0.0"
-__author__ = "Agent 3"
-__description__ = "Bulk Operations & Management for Type System"
+__author__ = "Agent 3 - Bulk Operations & Management"
+__description__ = "Comprehensive bulk operations system for type management"