--- a/src/torematrix/ui/components/element_list/advanced/__init__.py
+++ b/src/torematrix/ui/components/element_list/advanced/__init__.py
@@ -1,0 +1,18 @@
+"""
+Advanced Features for Hierarchical Element List
+
+This package provides advanced functionality including breadcrumb navigation,
+sorting, grouping, and bookmark systems for enhanced user experience.
+"""
+
+from .breadcrumb import BreadcrumbNavigator
+from .sorting import AdvancedSortingManager
+from .grouping import ElementGroupingManager
+from .bookmarks import BookmarkSystem
+
+__all__ = [
+    'BreadcrumbNavigator',
+    'AdvancedSortingManager', 
+    'ElementGroupingManager',
+    'BookmarkSystem'
+]