--- a/tests/unit/core/operations/type_management/__init__.py
+++ b/tests/unit/core/operations/type_management/__init__.py
@@ -1,0 +1,9 @@
+"""Test package for Type Management Operations
+
+Comprehensive test suite for bulk operations and management including:
+- Bulk type operations with threading and concurrency
+- Type conversions with data preservation
+- Progress tracking with real-time updates
+- Rollback and undo functionality
+- Performance optimization testing
+"""