… src/torematrix/core/operations/type_management/rollback.py
… src/torematrix/core/types/import_export.py
… src/torematrix/core/types/plugin_system.py
… src/torematrix/ui/components/element_list/advanced/bookmarks.py
… src/torematrix/ui/components/element_list/advanced/grouping.py
… src/torematrix/ui/components/element_list/advanced/sorting.py
… src/torematrix/ui/components/search/parallel.py
… tests/integration/property_panel/test_integration.py
… tests/unit/core/operations/type_management/test_bulk_operations.py
… tests/unit/core/operations/type_management/test_conversions.py
… tests/unit/core/operations/type_management/test_progress.py
… tests/unit/core/operations/type_management/test_rollback.py